+++ app/controllers/budgets/groups_controller.rb
@@ -1,6 +1,5 @@
 module Budgets
   class GroupsController < ApplicationController
-    include InvestmentFilters
     include FeatureFlags
     feature_flag :budgets
 
@@ -8,9 +7,6 @@ module Budgets
     before_action :load_group, only: [:show]
     authorize_resource :budget
     authorize_resource :group, class: "Budget::Group"
-
-    before_action :set_default_investment_filter, only: :show
-    has_filters investment_filters, only: [:show]
 
     def show
     end
