+++ app/views/budgets/groups/show.html.erb
@@ -1,26 +1,8 @@
-<% content_for :canonical do %>
-  <%= render "shared/canonical", href: budget_group_url(filter: @current_filter) %>
-<% end %>
-
 <main class="budget-group-show">
   <header>
     <%= back_link_to budget_path(@budget) %>
     <h1><%= t("budgets.groups.show.title") %></h1>
   </header>
-
-  <% if @current_filter == "unfeasible" %>
-    <div class="row margin-top">
-      <div class="small-12 column">
-        <h3><%= t("budgets.groups.show.unfeasible_title") %></h3>
-      </div>
-    </div>
-  <% elsif @current_filter == "unselected" %>
-    <div class="row margin-top">
-      <div class="small-12 column">
-        <h3><%= t("budgets.groups.show.unselected_title") %></h3>
-      </div>
-    </div>
-  <% end %>
 
   <div class="row margin">
     <div id="headings" class="small-12 medium-7 column select-district">
@@ -30,9 +12,8 @@
             <% slice.each do |heading| %>
               <span id="<%= dom_id(heading) %>"
                     class="<%= css_for_ballot_heading(heading) %>">
-                <%= link_to heading.name,
-                            budget_investments_path(heading_id: heading.id,
-                                                    filter: @current_filter) %><br>
+                <%= link_to heading.name, budget_investments_path(heading_id: heading.id) %>
+                <br>
               </span>
             <% end %>
           </div>
@@ -44,28 +25,4 @@
       <%= image_tag(image_path_for("map.jpg")) %>
     </div>
   </div>
-
-  <% if @budget.balloting_or_later? %>
-    <% unless @current_filter == "unfeasible" %>
-      <div class="row">
-        <div class="small-12 column">
-          <small>
-            <%= link_to t("budgets.groups.show.unfeasible"),
-                        budget_group_path(@budget, @group, filter: "unfeasible") %>
-          </small>
-        </div>
-      </div>
-    <% end %>
-
-    <% unless @current_filter == "unselected" %>
-      <div class="row">
-        <div class="small-12 column">
-          <small>
-            <%= link_to t("budgets.groups.show.unselected"),
-                        budget_group_path(@budget, @group, filter: "unselected") %>
-          </small>
-        </div>
-      </div>
-    <% end %>
-  <% end %>
 </main>
