+++ config/locales/es/budgets.yml
@@ -43,10 +43,6 @@ es:
     groups:
       show:
         title: Selecciona una partida
-        unfeasible_title: Proyectos de gasto inviables
-        unfeasible: Ver proyectos inviables
-        unselected_title: Proyectos no seleccionados para la votación final
-        unselected: Ver los proyectos no seleccionados para la votación final
     phase:
       drafting: Borrador (No visible para el público)
       informing: Información
